requirements.txt
Django==1.7.1
beautifulsoup4==4.3.2
blessings==1.6
cssselect==0.9.1
django-extensions==1.4.6
django-tastypie==0.12.1
enum34==0.9.23
facebook-sdk==0.4.0
lxml==3.3.4
mechanize==0.2.5
pymongo==2.7.2
python-dateutil==2.2
requests==2.2.1
six==1.6.1
wsgiref==0.1.2
unicodecsv==0.9.4
nltk==3.0.0